Core implementation code drop
I was rather undisciplined while implementing most of the internals and
didn't make any commits. So this is a bid of a code drop.

I still need to finish the last bits of the upgrade tarball generation
and then spend a significant amount of time documenting how this whole
thing works.