Massage final upgrade tarballs for convenience
These are a few extra tweaks that rebar provides so that the result of
applying an upgrade tarball results in something that looks like if you
had unpacked the release tarball. This helps scripts so they can pull
information from releases/start_erl.data and rely on having a consistent
version directory.